Removed InMemoryReplicationClient
